CORCI-277 check_python.sh bug fixes.
check_python.sh was using $GERRIT_PROJECT as the directory
to search for python code.

It is now using $PROJECT_REPO.  If PROJECT_REPO was not defined,
then the $GERRIT_PROJECT is used to find a directory to set
PROJECT_REPO to.

The code to find if the check_module.sh was fixed to only look in
the $PROJECT_REPO directory.

The output of check_module.sh needs to be filtered to only include the
files that are in the review.

Change-Id: I3c88735135543ee351da72ae08095793cd122311
Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>
Reviewed-on: https://review.whamcloud.com/32057
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>