CORCI-579 Code review shouldn't add a positive review
Our projects require two positive reviews from developers
to land.  We shouldn't be allowing the checkpatch code
review to be one of those.

So just have checkpatch either provide a negative review
or a neutral (COMMENT) one.

If checkpatch does apply a negative review, gatekeepers can
(and will have to) dismiss the checkpatch negative reivew
for the patch to land.  This is positive acknowledgement
from the gatekeeper that they have seen the checkpatch
comments and have allowed it to land anyway.

Change-Id: I3295a185216646a4646a60f921845dd2b586f51a
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: https://review.hpdd.intel.com/34155
Tested-by: Jenkins
Reviewed-by: John Malmberg <john.e.malmberg@intel.com>