CORCI-611 Truncate GitHub code review comments at 64K
GitHub appears to have a limit on PR comments of 64K so just truncate
review comments at 64K with a message about the truncation.

We could of course just split it into multiple 64K comments, but once
a code review comment has reached 64K it's usefulness is probably
limited and the developer will likely want to go back to their
development environment to resolve the mass of code review issues if
they need to be dealt with at all (i.e. they could be advisory-only -
in unmodified lines).

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>