CORCI-694 Don't post a review when checkpatch is successful
Since we don't want a comment when it's successful, but cannot post
anything other than a +1 review without a comment, simply don't post a
review when there is no errors.

Also, don't show incidental linting for daos_api.py.  This will be
resolved in the future when the cleaning up of that file has been
completed.
- this (partially) reverts commit 8b33e1faacaaadb0f3ad7eb6634d9c04ca847de0.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>