CORCI-27 test: Add code to give suggestions to github.
Add support for giving suggestions back to patches in github.
If a script with the right name exits in the source then call it
to patch the src tree in place, then run "git diff" to parse these
changes and report them back to github as suggestions.

Port github_checkpatch to python3, which requires docker/package
changes.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>